Change Color
I also want to add the function for testing purposes to change the
color of the ball in the cannon. I am on vacation and will work on it
when I can, but there is some code in there already to start this
function. Just adding this code and comments in case someone works on
it while I am away